% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_discards.R
\name{get_discard_ratio_per_category}
\alias{get_discard_ratio_per_category}
\title{Get the discard ratio for each category}
\usage{
get_discard_ratio_per_category(so)
}
\arguments{
\item{so}{}
}
\description{
Get the discard ratio for each category
}
