% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_discards.R
\name{group_gears_add_percent_discard}
\alias{group_gears_add_percent_discard}
\title{Make gear groupings and get the percent catch}
\usage{
group_gears_add_percent_discard(so, groups)
}
\arguments{
\item{so}{Output from read_StockOverview}

\item{groups}{}
}
\description{
Make gear groupings and get the percent catch
}
