% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_discards.R
\name{raise_discards}
\alias{raise_discards}
\title{Read stock overview files and raise dicards according to a grouping scheme}
\usage{
raise_discards(
  fn,
  groups,
  nodiscardgroups = c("NO", "MIS"),
  allgroups = NULL,
  threshold = 1e+10,
  v = FALSE
)
}
\arguments{
\item{fn}{StockOverview.txt filenames}

\item{groups}{Grouppings of gears/areas/countries to do the raising}

\item{nodiscardgroups}{Grouips assumed to have no discards}

\item{allgroups}{Groups that use the overall mean discard ratio for raising discards}

\item{threshold}{Remove metiers with discard rates above this threshold from the group means}

\item{v}{Should I be verbose?}
}
\description{
Read stock overview files and raise dicards according to a grouping scheme
}
