% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_discards.R
\name{read_StockOverview}
\alias{read_StockOverview}
\title{Reads stockoverview files}
\usage{
read_StockOverview(fn, v = FALSE)
}
\arguments{
\item{fn}{filename}

\item{v}{if TRUE, show messages}
}
\description{
Reads stockoverview files
}
